Share `Device` logic among Gecko and Servo (#330)
Adds a new file `style/device/mod.rs` with the `Device` logic that is
common for Gecko and Servo.

The Gecko-specific logic for `Device` is moved into
`style/device/gecko.rs`, and the previous `style/gecko/media_queries.rs`
is removed.

The Servo-specific logic for `Device` is moved into
`style/device/servo.rs`. The previous `servo/media_queries.rs` also had
logic for media features, which is now moved into
`style/servo/media_features.rs` for consistency with Gecko.

Servo PR: https://github.com/servo/servo/pull/43146
