+++ style/invalidation/stylesheets.rs
@@ -10,10 +10,10 @@
 use crate::context::QuirksMode;
 use crate::data::ElementData;
 use crate::derives::*;
+use crate::device::Device;
 use crate::dom::{TDocument, TElement, TNode};
 use crate::invalidation::element::element_wrapper::{ElementSnapshot, ElementWrapper};
 use crate::invalidation::element::restyle_hints::RestyleHint;
-use crate::media_queries::Device;
 use crate::selector_map::PrecomputedHashSet;
 use crate::selector_parser::{SelectorImpl, Snapshot, SnapshotMap};
 use crate::shared_lock::SharedRwLockReadGuard;
