+++ style/media_queries/media_list.rs
@@ -6,9 +6,10 @@
 //!
 //! https://drafts.csswg.org/mediaqueries/#typedef-media-query-list
 
-use super::{Device, MediaQuery, Qualifier};
+use super::{MediaQuery, Qualifier};
 use crate::context::QuirksMode;
 use crate::derives::*;
+use crate::device::Device;
 use crate::error_reporting::ContextualParseError;
 use crate::parser::ParserContext;
 use crate::stylesheets::CustomMediaEvaluator;
