+++ style/servo/mod.rs
@@ -10,7 +10,7 @@ pub mod animation;
 #[allow(missing_docs)] // TODO.
 pub mod attr;
 mod encoding_support;
-pub mod media_queries;
+pub mod media_features;
 pub mod restyle_damage;
 pub mod selector_parser;
 mod shadow_parts;
