+++ style/stylesheet_set.rs
@@ -5,8 +5,8 @@
 //! A centralized set of stylesheets for a document.
 
 use crate::derives::*;
+use crate::device::Device;
 use crate::invalidation::stylesheets::{RuleChangeKind, StylesheetInvalidationSet};
-use crate::media_queries::Device;
 use crate::shared_lock::SharedRwLockReadGuard;
 use crate::stylesheets::{
     CssRule, CssRuleRef, CustomMediaMap, Origin, OriginSet, PerOrigin, StylesheetInDocument,
